changed scouterID to INT

+++ SQL/dev1/pit-01.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS dev1.pit (
         team VARCHAR(10) NOT NULL,
         eventID INT NOT NULL,
         scoutingStatus TINYINT DEFAULT 0,
-        scouterID TINYINT NULL,
+        scouterID INT NULL,
         robotLength TINYINT NULL,
         robotWidth TINYINT NULL,
         robotHeight TINYINT NULL,
